## Working Commit ##
This commit is the **working** revision of the refactoring of
unit-related sources into src/unit/ from src as well as the cleanup of
several useless namespaces and the refactoring of unit_ability_list into
its own source/header files from unit.hpp.unit/abilities.cpp.

This is the first step in cleaning up and refactoring unit.